+++ src/Facades/Audio.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * @method static bool play(string $url, array $options = [])
+ * @method static bool playStream(string $mountpoint, array $options = [])
  * @method static bool load(string $url, array $options = [])
  * @method static bool pause()
  * @method static bool resume()
@@ -17,6 +18,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static float|null getCurrentPosition()
  * @method static array getState()
  * @method static bool setMetadata(array $metadata)
+ * @method static bool updateStreamMetadata(array $metadata)
  * @method static array drainEvents()
  * @method static bool setPlaylist(array $tracks, bool $autoPlay = true, int $startIndex = 0)
  * @method static bool next()